refactor: Position, File, Rank 패키지 이동
