+++ src/main/java/chess/domain/order/Order.java
@@ -1,6 +1,6 @@
 package chess.domain.order;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.Arrays;
