+++ src/main/java/chess/view/OutputView.java
@@ -1,9 +1,9 @@
 package chess.view;
 
-import chess.File;
-import chess.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Rank;
 import chess.domain.Board;
-import chess.domain.Position;
+import chess.domain.position.Position;
 import chess.domain.piece.Piece;
 
 import java.util.Map;
