+++ src/test/java/chess/domain/order/OrderTest.java
@@ -1,6 +1,6 @@
 package chess.domain.order;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
